Remove StyledText from public API for now
The StyledText element is still experimental API, so keep all the StyledText
apis as private.

Also make @markdown as experimental
