+++ api/cpp/include/slint_string.h
@@ -223,6 +223,8 @@ private:
     void *inner; // opaque
 };
 
+namespace private_api {
+
 /// Styled text that has been parsed and seperated into paragraphs
 struct StyledText
 {
@@ -259,7 +261,6 @@ public:
 private:
     void *inner;
 };
-namespace private_api {
 
 template<typename T>
 inline cbindgen_private::Slice<T> make_slice(const T *ptr, size_t len)
@@ -282,5 +283,4 @@ inline cbindgen_private::Slice<uint8_t> string_to_slice(std::string_view str)
     return make_slice(reinterpret_cast<const uint8_t *>(str.data()), str.size());
 }
 }
-
 }
