+++ api/cpp/lib.rs
@@ -11,8 +11,8 @@ extern crate std;
 use alloc::rc::Rc;
 use core::ffi::c_void;
 use i_slint_core::SharedString;
-use i_slint_core::api::StyledText;
 use i_slint_core::items::OperatingSystemType;
+use i_slint_core::styled_text::StyledText;
 use i_slint_core::window::{WindowAdapter, ffi::WindowAdapterRcOpaque};
 
 pub mod platform;
@@ -242,11 +242,11 @@ pub extern "C" fn slint_detect_operating_system() -> OperatingSystemType {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn slint_escape_markdown(text: &mut SharedString) -> &SharedString {
-    *text = i_slint_core::escape_markdown(text).into();
+    *text = i_slint_core::styled_text::escape_markdown(text).into();
     text
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn slint_parse_markdown(text: &SharedString, out: &mut StyledText) {
-    *out = i_slint_core::parse_markdown(text);
+    *out = i_slint_core::styled_text::parse_markdown(text);
 }
