+++ internal/compiler/generator/cpp.rs
@@ -558,7 +558,7 @@ impl CppType for Type {
             Type::LayoutCache => Some("slint::SharedVector<float>".into()),
             Type::ArrayOfU16 => Some("slint::SharedVector<uint16_t>".into()),
             Type::Easing => Some("slint::cbindgen_private::EasingCurve".into()),
-            Type::StyledText => Some("slint::StyledText".into()),
+            Type::StyledText => Some("slint::private_api::StyledText".into()),
             _ => None,
         }
     }
